--- jiyoontrip/src/app/components/boards-list/pagination/queires.ts
@@ -1,32 +1,0 @@
-import { gql } from "@apollo/client";
-
-export const FERTCH_BOARDS_COUNT = gql`
-  query {
-    fetchBoardsCount
-  }
-`;
-
-export const FETCH_BOARDS = gql`
-  query fetchBoards($page: Int) {
-    fetchBoards(page: $page) {
-      _id
-      writer
-      title
-      contents
-      youtubeUrl
-      likeCount
-      dislikeCount
-      images
-      boardAddress {
-        _id
-        zipcode
-        address
-        addressDetail
-        createdAt
-        updatedAt
-      }
-      createdAt
-      updatedAt
-    }
-  }
-`;
